requirements.txt
pytest==7.1.2
requests==2.28.1
urllib3==1.26.10
selenium==4.3.0
allure-pytest==2.9.45